+++ app/javascript/pages/PageTimeline.vue
@@ -3,14 +3,10 @@
         <v-row>
             <v-col>
                 <div class="pb-10">
-                    <div class="mb-5" v-if="$store.getters['auth/currentUser']">
-                        <micropost-form @created="createMicropost"></micropost-form>
-                    </div>
                     <timeline-list v-if="isExistMicroposts" :microposts="microposts" class="mb-5"></timeline-list>
                     <div class="text-center" v-else>
                         一件もありません
                     </div>
-
                     <template v-if="pagingMeta">
                         <div class="text-center">
                             <v-pagination
@@ -27,7 +23,6 @@
     </v-container>
 </template>
 <script>
-    import MicropostForm from '@/components/MicropostForm'
     import TimelineList from '@/components/TimelineList'
     import axios from "axios";
     export default {
@@ -39,7 +34,6 @@
             }
         },
         components: {
-            MicropostForm,
             TimelineList
         },
         computed: {
@@ -55,15 +49,6 @@
                 const res = await axios.get(`/api/microposts`, { params: { page: this.currentPage } })
                 this.microposts = res.data.microposts
                 this.pagingMeta = res.data.meta
-            },
-            async createMicropost(micropostContent) {
-                const micropostParams = {
-                    micropost: {
-                        content: micropostContent
-                    }
-                }
-                const res = await axios.post(`/api/microposts`, micropostParams)
-                this.microposts =[...[res.data.micropost], ...this.microposts]
             },
             paging(page) {
                 this.currentPage = page
