+++ app/javascript/router/index.js
@@ -4,6 +4,7 @@ Vue.use(VueRouter)
 import PageUserRegister from '@/pages/PageUserRegister'
 import PageUserLogin from '@/pages/PageUserLogin'
 import PageTimeline from '@/pages/PageTimeline'
+import PageMicropostPost from '@/pages/PageMicropostPost'
 import PageMicropostDetail from '@/pages/PageMicropostDetail'
 import PageProfile from '@/pages/PageProfile'
 import PageUsers from '@/pages/PageUsers'
@@ -14,6 +15,7 @@ const router = new VueRouter({
         { path: '/login', component: PageUserLogin, name: 'user-login' },
         { path: '/', component: PageTimeline, name: 'timeline' },
         { path: '/microposts', component: PageTimeline },
+        { path: '/micropost', component: PageMicropostPost, name: 'micropost-post' },
         { path: '/microposts/:id', component: PageMicropostDetail, name: 'micropost-detail' },
         { path: '/profile', component: PageProfile, name: 'my-profile' },
         { path: '/users/:id', component: PageProfile, name: 'user-profile' },
