+++ app/models/user.rb
@@ -14,7 +14,7 @@ class User < ApplicationRecord
   scope :by_tag, ->(tag_ids) { joins(:user_tags).where(user_tags: { tag_id: tag_ids }) }
 
   def save_with_tags!(tag_names:)
-    return save! if tag_names.nil? # blank?だと削除ができない
+    return save! if tag_names.nil? 
 
     ActiveRecord::Base.transaction do
       self.tags = tag_names.map { |name| Tag.find_or_initialize_by(name: name) }
